requirements.txt
langchain==0.0.184
PyPDF2==3.0.1
PyPDF==3.17.0
python-dotenv==1.0.0
streamlit==1.28.0
openai==0.27.6
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
qdrant_client==1.6.4
# uncomment to use huggingface llms
# huggingface-hub==0.14.1

# uncomment to use instructor embeddings
# InstructorEmbedding==1.0.1
# sentence-transformers==2.2.2
